Rollup branch: Initial commit
Seems to be a lot smoother than Webpack 2.

The hierarchy is a little wonky with all of the `index.js` modules, but
the library is running.

Examples are building with `npm run examples`, which also starts a
server at `localhost:10001`.

The new top level directory in the source tree can stay or go, but it
made it slightly easier to get the build running.