New query logic, data structures, babel features
During real-world testing, it was discovered that the query engine was
only returning exact matches for component keys; ie, systems would only
get entities that had an identical set of components to the query, and
would not get those with a superset of components in their
`componentKey`. `EntityManager.query(componentsKey)` now returns a `Set`
of entities associate with a superset of components.

The datastructures for `componentEntityMap` and `entityComponents` have
been fixed so that an unnecessary array index is not necessary to look
up components.

Profiling revealed some massive slowdowns due to Babel transpiled ES6/7
features. I had assumed that `babel-preset-env` would take care of these
details, but it turns out the recent FF and Chrome support nearly all of
the features used natively, and the `babel-preset-env` database was not
preventing these from being emulated. A growing blacklist of plugins has
been added.

* New `EntityManager.query(componentsKey)`, `entityFeed` now yielding
  expected entities to systems.
* Fixed data structures to remove accidental redundant arrays.
* Excluded Babel plugins where native support was discovered.
* Extended builtin `Set` with `isSuperSet()`.
* Used `var` to work around inexplicable v8 inability to work with `let`
  variables and compound operators (e.g., `let foo = 0; foo += 1;`.
  Added TODO to research adding Babel plugins for performance
  optimizations.
* Added `reqId` for `World.stop(reqId)`, so that loops calling
  `cancelAnimationFrame(reqId) works.

NOTE Mocha tests are certainly broken due to API changes and no
attention being paid to `node` polyfills. Fix soon...